Le pongo tipo de variable DOUBLE

+++ TP1/conversion.c
@@ -1,9 +1,9 @@
 #include "conversion.h"
 
-float fahrenheitToCelsius (float fahr){
+double fahrenheitToCelsius (double fahr){
     return (5.0 / 9.0) * (fahr - 32.0);
 }
 
-float celsiusToFahrenheit(float celsius){
+double celsiusToFahrenheit(double celsius){
     return ((5.0 / 9.0) * celsius) + 32.0;
 }
